// src/openapi/client/models/PlaceTranslationVisible.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { ChoiceBase_LanguagesChoices_ } from './ChoiceBase_LanguagesChoices_';
export type PlaceTranslationVisible = {
    id: number;
    name: string;
    fullName: string;
    locale: ChoiceBase_LanguagesChoices_;
};